Introduce "register subscription" event, move declarative links decorator to addon-links
