+++ addons/links/src/index.js
@@ -3,7 +3,7 @@ export const EVENT_ID = `${ADDON_ID}/link-event`;
 export const REQUEST_HREF_EVENT_ID = `${ADDON_ID}/request-href-event`;
 export const RECEIVE_HREF_EVENT_ID = `${ADDON_ID}/receive-href-event`;
 
-export { linkTo, hrefTo } from './preview';
+export { linkTo, hrefTo, withLinks } from './preview';
 
 let hasWarned = false;
