+++ lib/channel-postmessage/src/index.js
@@ -32,7 +32,7 @@ export class PostmsgTransport {
     }
     const data = stringify({ key: KEY, event });
     iframeWindow.postMessage(data, '*');
-    window.postMessage(data, '*');
+    this._handler(event);
     return Promise.resolve(null);
   }
