+++ lib/core-events/index.js
@@ -8,4 +8,5 @@ module.exports = {
   APPLY_SHORTCUT: 'applyShortcut',
   STORY_ADDED: 'storyAdded',
   FORCE_RE_RENDER: 'forceReRender',
+  REGISTER_SUBSCRIPTION: 'registerSubscription',
 };
