+++ lib/core/src/client/preview/index.js
@@ -5,7 +5,6 @@ import ConfigApi from './config_api';
 import StoryStore from './story_store';
 import reducer from './reducer';
 import syncUrlWithStore from './syncUrlWithStore';
-import subscriptionsStore from './subscriptions_store';
 
 export default {
   start,
@@ -15,5 +14,4 @@ export default {
   StoryStore,
   reducer,
   syncUrlWithStore,
-  subscriptionsStore,
 };
