+++ lib/core/src/client/preview/start.js
@@ -89,6 +89,7 @@ export default function start(render, { decorateStory } = {}) {
     channel.on(Events.SET_CURRENT_STORY, data => {
       reduxStore.dispatch(Actions.selectStory(data.kind, data.story));
     });
+    channel.on(Events.REGISTER_SUBSCRIPTION, subscriptionsStore.register);
     addons.setChannel(channel);
     Object.assign(context, { channel });
