+++ lib/core/src/client/preview/subscriptions_store.js
@@ -1,5 +1,3 @@
-import global from 'global';
-
 export const createSubscriptionsStore = () => {
   const subscripions = new Map();
 
@@ -33,9 +31,4 @@ export const createSubscriptionsStore = () => {
   };
 };
 
-// Enforce global singleton
-const KEY = '__STORYBOOK_SUBSCRIPTIONS';
-if (!global[KEY]) {
-  global[KEY] = createSubscriptionsStore();
-}
-export default global[KEY];
+export default createSubscriptionsStore();
